+++ CHANGELOG.md
@@ -14,6 +14,8 @@ See [STATUS.md](server/STATUS.md) to learn more about which features will remain
 - Fix `--script` flag
 - Rename `setup-env` to `generate-dotenv` and build it from clap #599
 - Remove `remove_previous_search` and `asset_url` options
+- Add `Storelike::post_resource`, which allows plugins to parse HTTP POST requests #592
+- Move Server-Timing header to crate `simple-server-timing-header`
 
 ## [v0.34.1] - 2023-02-11